Fix Hidden Window in Window Stack & Play/Stop Music
I want to use the window when the break is over, but it keeps going back to its original place and does not get moved to the top of the
window stack.

Also changed Play/Pause to Play/Stop in order to work with my Di.fm
streaming playlist. Stopping will prevent the player from timing out.